+++ netpyne/simFuncs.py
@@ -25,8 +25,7 @@ from specs import Dict, ODict
 from collections import OrderedDict
 from neuron import h, init # Import NEURON
 import sim, specs
-import tests
-from tests.tests import *
+
 ###############################################################################
 # initialize variables and MPI
 ###############################################################################
@@ -37,11 +36,14 @@ def initialize (netParams = None, simConfig = None, net = None):
         print('Error: seems like the sim.initialize() arguments are in the wrong order, try initialize(netParams, simConfig)')
         sys.exit()
 
-    # if sim config
     if simConfig.checkErrors: # whether to validate the input parameters
-        netPyneTestObj = NetPyneTestObj(simConfig.checkErrorsVerbose)
-        netPyneTestObj.netParams = netParams
-        netPyneTestObj.runTests()
+        simTestObj = sim.SimTestObj(simConfig.checkErrorsVerbose)
+        simTestObj.netParams = netParams
+        simTestObj.runTests()
+
+    # for testing validation
+    # if simConfig.exitOnError:
+    #sys.exit()
 
     sim.simData = Dict()  # used to store output simulation data (spikes etc)
     sim.fih = []  # list of func init handlers
@@ -788,7 +790,7 @@ def preRun ():
     # set initial v of cells
     sim.fih = []
     for cell in sim.net.cells:
-       sim.fih.append(h.FInitializeHandler(cell.initV))
+       sim.fih.append(h.FInitializeHandler(0, cell.initV))
 
     # cvode variables
     if not getattr(h, 'cvode', None):