+++ netpyne/tests/validate_tests.py
@@ -7,7 +7,7 @@ Contributors: mitra.sidddhartha@gmail.com
 """
 
 from tests import *
-import specs
+import netpyne.specs as specs
 
 class ParamsObj():
 
@@ -38,150 +38,192 @@ class RunNetPyneTests():
             self.paramsMap["stimSource"] = {}
             self.paramsMap["stimTarget"] = {}
 
-            self.paramsMap["pop"]["cellModelTest"] = []
+            # self.paramsMap["pop"]["cellModelTest"] = []
+            #
+            # cellModelParams = ParamsObj()
+            # cellModelParams.netParams.popParams['validCellModelParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'numCells': 50}     # add dict with params for this pop
+            # self.paramsMap["pop"]["cellModelTest"].append(cellModelParams)
+            #
+            # cellModelParams = ParamsObj()
+            # cellModelParams.netParams.popParams['invalidCellModelParams'] = {'cellType': 'PYR', 'numCells': 50}     # add dict with params for this pop
+            # self.paramsMap["pop"]["cellModelTest"].append(cellModelParams)
+            #
+            # self.paramsMap["pop"]["volumeParamsTest"] = []
+            #
+            # volumeParams = ParamsObj()
+            # volumeParams.netParams.popParams['validVolumeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50}     # add dict with params for this pop
+            # self.paramsMap["pop"]["volumeParamsTest"].append(volumeParams)
+            #
+            # volumeParams = ParamsObj()
+            # volumeParams.netParams.popParams['invalidVolumeParams'] = {'cellType': 'PYR', 'cellModel': 'HH'}     # add dict with params for this pop
+            # self.paramsMap["pop"]["volumeParamsTest"].append(volumeParams)
+            #
+            # self.paramsMap["pop"]["xNormRangeParamsTest"] = []
+            #
+            # params = ParamsObj()
+            # params.netParams.popParams['validxNormRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'xnormRange' : [0.6,0.9]}     # add dict with params for this pop
+            # self.paramsMap["pop"]["xNormRangeParamsTest"].append(params)
+            #
+            # params = ParamsObj()
+            # params.netParams.popParams['invalidxNormRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'xnormRange' : 0.6}     # add dict with params for this pop
+            # self.paramsMap["pop"]["xNormRangeParamsTest"].append(params)
+            #
+            # params = ParamsObj()
+            # params.netParams.popParams['invalidxNormRangeParams1'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'xnormRange' : [6,10]}     # add dict with params for this pop
+            # self.paramsMap["pop"]["xNormRangeParamsTest"].append(params)
+            #
+            # self.paramsMap["pop"]["yNormRangeParamsTest"] = []
+            #
+            # params = ParamsObj()
+            # params.netParams.popParams['validyNormRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'ynormRange' : [0.6,0.9]}     # add dict with params for this pop
+            # self.paramsMap["pop"]["yNormRangeParamsTest"].append(params)
+            #
+            # params = ParamsObj()
+            # params.netParams.popParams['invalidyNormRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'ynormRange' : 0.6}     # add dict with params for this pop
+            # self.paramsMap["pop"]["yNormRangeParamsTest"].append(params)
+            #
+            # params = ParamsObj()
+            # params.netParams.popParams['invalidyNormRangeParams1'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'ynormRange' : [6,10]}     # add dict with params for this pop
+            # self.paramsMap["pop"]["yNormRangeParamsTest"].append(params)
+            #
+            # self.paramsMap["pop"]["zNormRangeParamsTest"] = []
+            #
+            # params = ParamsObj()
+            # params.netParams.popParams['validzNormRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'znormRange' : [0.6,0.9]}     # add dict with params for this pop
+            # self.paramsMap["pop"]["zNormRangeParamsTest"].append(params)
+            #
+            # params = ParamsObj()
+            # params.netParams.popParams['invalidzNormRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'znormRange' : 0.6}     # add dict with params for this pop
+            # self.paramsMap["pop"]["zNormRangeParamsTest"].append(params)
+            #
+            # params = ParamsObj()
+            # params.netParams.popParams['invalidzNormRangeParams1'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'znormRange' : [6,10]}     # add dict with params for this pop
+            # self.paramsMap["pop"]["zNormRangeParamsTest"].append(params)
+            #
+            # self.paramsMap["pop"]["zNormRangeParamsTest"] = []
+            #
+            # params = ParamsObj()
+            # params.netParams.popParams['validzNormRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'znormRange' : [0.6,0.9]}     # add dict with params for this pop
+            # self.paramsMap["pop"]["zNormRangeParamsTest"].append(params)
+            #
+            # params = ParamsObj()
+            # params.netParams.popParams['invalidzNormRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'znormRange' : 0.6}     # add dict with params for this pop
+            # self.paramsMap["pop"]["zNormRangeParamsTest"].append(params)
+            #
+            # params = ParamsObj()
+            # params.netParams.popParams['invalidzNormRangeParams1'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'znormRange' : [6,10]}     # add dict with params for this pop
+            # self.paramsMap["pop"]["zNormRangeParamsTest"].append(params)
+            #
+            # self.paramsMap["pop"]["xRangeParamsTest"] = []
+            #
+            # params = ParamsObj()
+            # params.netParams.sizeX = 70   # max size for network
+            # params.netParams.popParams['validxRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'xRange' : [30,60]}     # add dict with params for this pop
+            # self.paramsMap["pop"]["xRangeParamsTest"].append(params)
+            #
+            # params = ParamsObj()
+            # params.netParams.sizeX = 70   # max size for network
+            # params.netParams.popParams['invalidxRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'xRange' : [40,90]}     # add dict with params for this pop
+            # self.paramsMap["pop"]["xRangeParamsTest"].append(params)
+            #
+            # self.paramsMap["pop"]["yRangeParamsTest"] = []
+            #
+            # params = ParamsObj()
+            # params.netParams.sizeY = 70   # max size for network
+            # params.netParams.popParams['validyRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'yRange' : [30,60]}     # add dict with params for this pop
+            # self.paramsMap["pop"]["yRangeParamsTest"].append(params)
+            #
+            # params = ParamsObj()
+            # params.netParams.sizeY = 70   # max size for network
+            # params.netParams.popParams['invalidyRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'yRange' : [40,90]}     # add dict with params for this pop
+            # self.paramsMap["pop"]["yRangeParamsTest"].append(params)
+            #
+            # self.paramsMap["pop"]["zRangeParamsTest"] = []
+            #
+            # params = ParamsObj()
+            # params.netParams.sizeZ = 70   # max size for network
+            # params.netParams.popParams['validzRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'zRange' : [30,60]}     # add dict with params for this pop
+            # self.paramsMap["pop"]["zRangeParamsTest"].append(params)
+            #
+            # params = ParamsObj()
+            # params.netParams.sizeZ = 70   # max size for network
+            # params.netParams.popParams['invalidzRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'zRange' : [40,90]}     # add dict with params for this pop
+            # self.paramsMap["pop"]["zRangeParamsTest"].append(params)
 
-            cellModelParams = ParamsObj()
-            cellModelParams.netParams.popParams['validCellModelParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'numCells': 50}     # add dict with params for this pop
-            self.paramsMap["pop"]["cellModelTest"].append(cellModelParams)
-
-            cellModelParams = ParamsObj()
-            cellModelParams.netParams.popParams['invalidCellModelParams'] = {'cellType': 'PYR', 'numCells': 50}     # add dict with params for this pop
-            self.paramsMap["pop"]["cellModelTest"].append(cellModelParams)
-
-            self.paramsMap["pop"]["volumeParamsTest"] = []
-
-            volumeParams = ParamsObj()
-            volumeParams.netParams.popParams['validVolumeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50}     # add dict with params for this pop
-            self.paramsMap["pop"]["volumeParamsTest"].append(volumeParams)
-
-            volumeParams = ParamsObj()
-            volumeParams.netParams.popParams['invalidVolumeParams'] = {'cellType': 'PYR', 'cellModel': 'HH'}     # add dict with params for this pop
-            self.paramsMap["pop"]["volumeParamsTest"].append(volumeParams)
-
-            self.paramsMap["pop"]["xNormRangeParamsTest"] = []
+            self.paramsMap["pop"]["popStimParamsTest"] = []
+            
+            params = ParamsObj()
+            params.netParams.popParams['validPopStimParams1'] = {'cellModel': 'IntFire2', 'taum': 100, 'noise': 0.5, 'numCells': 100}  # Intfire2
+            self.paramsMap["pop"]["popStimParamsTest"].append(params)
 
             params = ParamsObj()
-            params.netParams.popParams['validxNormRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'xnormRange' : [0.6,0.9]}     # add dict with params for this pop
-            self.paramsMap["pop"]["xNormRangeParamsTest"].append(params)
+            params.netParams.popParams['validPopStimParams2'] = {'cellModel': 'NetStim', 'rate': 100, 'noise': 0.5, 'numCells': 100}  # NetsStim
+            self.paramsMap["pop"]["popStimParamsTest"].append(params)
 
             params = ParamsObj()
-            params.netParams.popParams['invalidxNormRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'xnormRange' : 0.6}     # add dict with params for this pop
-            self.paramsMap["pop"]["xNormRangeParamsTest"].append(params)
+            # create custom list of spike times
+            spkTimes = range(0,1000,20) + [138, 155,270]
+            # create list of pulses (each item is a dict with pulse params)
+            pulses = [{'start': 10, 'end': 100, 'rate': 200, 'noise': 0.5},{'start': 400, 'end': 500, 'rate': 1, 'noise': 0.0}]
+            params.netParams.popParams['validPopStimParams3'] = {'cellModel': 'VecStim', 'numCells': 100, 'spkTimes': spkTimes, 'pulses': pulses}  # VecStim with spike times
+            self.paramsMap["pop"]["popStimParamsTest"].append(params)
 
             params = ParamsObj()
-            params.netParams.popParams['invalidxNormRangeParams1'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'xnormRange' : [6,10]}     # add dict with params for this pop
-            self.paramsMap["pop"]["xNormRangeParamsTest"].append(params)
-
-            self.paramsMap["pop"]["yNormRangeParamsTest"] = []
+            params.netParams.popParams['invalidPopStimParams1'] = {'cellModel': 'IntFire2', 'taum': 100, 'noise': 2, 'numCells': 100}  # Intfire2
+            self.paramsMap["pop"]["popStimParamsTest"].append(params)
 
             params = ParamsObj()
-            params.netParams.popParams['validyNormRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'ynormRange' : [0.6,0.9]}     # add dict with params for this pop
-            self.paramsMap["pop"]["yNormRangeParamsTest"].append(params)
+            params.netParams.popParams['invalidPopStimParams2'] = {'cellModel': 'NetStim', 'rate': '2', 'noise': 0.5, 'numCells': 100}  # NetsStim
+            self.paramsMap["pop"]["popStimParamsTest"].append(params)
 
             params = ParamsObj()
-            params.netParams.popParams['invalidyNormRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'ynormRange' : 0.6}     # add dict with params for this pop
-            self.paramsMap["pop"]["yNormRangeParamsTest"].append(params)
+            # create custom list of spike times
+            spkTimes = 1000
+            # create list of pulses (each item is a dict with pulse params)
+            pulses = [{'start': 10, 'end': 100, 'rate': 200, 'noise': 0.5},{'start': 400, 'end': 500, 'rate': 1, 'noise': 0.0}]
+            params.netParams.popParams['invalidPopStimParams3'] = {'cellModel': 'VecStim', 'numCells': 100, 'spkTimes': spkTimes, 'pulses': pulses}  # VecStim with spike times
+            self.paramsMap["pop"]["popStimParamsTest"].append(params)
 
             params = ParamsObj()
-            params.netParams.popParams['invalidyNormRangeParams1'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'ynormRange' : [6,10]}     # add dict with params for this pop
-            self.paramsMap["pop"]["yNormRangeParamsTest"].append(params)
+            # create custom list of spike times
+            spkTimes = range(0,1000,20) + [138, 155,270]
+            # create list of pulses (each item is a dict with pulse params)
+            pulses = [{'start': 10, 'end ': 100, 'rate': 200, 'noise': 0.5}, {'start': 400, 'end': 500, 'rate': 1, 'noise': 0.0}]
+            params.netParams.popParams['invalidPopStimParams4'] = {'cellModel': 'VecStim', 'numCells': 100, 'spkTimes': spkTimes, 'pulses': pulses}  # VecStim with spike times
+            self.paramsMap["pop"]["popStimParamsTest"].append(params)
 
-            self.paramsMap["pop"]["zNormRangeParamsTest"] = []
-
-            params = ParamsObj()
-            params.netParams.popParams['validzNormRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'znormRange' : [0.6,0.9]}     # add dict with params for this pop
-            self.paramsMap["pop"]["zNormRangeParamsTest"].append(params)
-
-            params = ParamsObj()
-            params.netParams.popParams['invalidzNormRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'znormRange' : 0.6}     # add dict with params for this pop
-            self.paramsMap["pop"]["zNormRangeParamsTest"].append(params)
-
-            params = ParamsObj()
-            params.netParams.popParams['invalidzNormRangeParams1'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'znormRange' : [6,10]}     # add dict with params for this pop
-            self.paramsMap["pop"]["zNormRangeParamsTest"].append(params)
-
-            self.paramsMap["pop"]["zNormRangeParamsTest"] = []
-
-            params = ParamsObj()
-            params.netParams.popParams['validzNormRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'znormRange' : [0.6,0.9]}     # add dict with params for this pop
-            self.paramsMap["pop"]["zNormRangeParamsTest"].append(params)
-
-            params = ParamsObj()
-            params.netParams.popParams['invalidzNormRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'znormRange' : 0.6}     # add dict with params for this pop
-            self.paramsMap["pop"]["zNormRangeParamsTest"].append(params)
-
-            params = ParamsObj()
-            params.netParams.popParams['invalidzNormRangeParams1'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'znormRange' : [6,10]}     # add dict with params for this pop
-            self.paramsMap["pop"]["zNormRangeParamsTest"].append(params)
-
-            self.paramsMap["pop"]["xRangeParamsTest"] = []
-
-            params = ParamsObj()
-            params.netParams.sizeX = 70   # max size for network
-            params.netParams.popParams['validxRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'xRange' : [30,60]}     # add dict with params for this pop
-            self.paramsMap["pop"]["xRangeParamsTest"].append(params)
-
-            params = ParamsObj()
-            params.netParams.sizeX = 70   # max size for network
-            params.netParams.popParams['invalidxRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'xRange' : [40,90]}     # add dict with params for this pop
-            self.paramsMap["pop"]["xRangeParamsTest"].append(params)
-
-            self.paramsMap["pop"]["yRangeParamsTest"] = []
-
-            params = ParamsObj()
-            params.netParams.sizeY = 70   # max size for network
-            params.netParams.popParams['validyRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'yRange' : [30,60]}     # add dict with params for this pop
-            self.paramsMap["pop"]["yRangeParamsTest"].append(params)
-
-            params = ParamsObj()
-            params.netParams.sizeY = 70   # max size for network
-            params.netParams.popParams['invalidyRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'yRange' : [40,90]}     # add dict with params for this pop
-            self.paramsMap["pop"]["yRangeParamsTest"].append(params)
-
-            self.paramsMap["pop"]["zRangeParamsTest"] = []
-
-            params = ParamsObj()
-            params.netParams.sizeZ = 70   # max size for network
-            params.netParams.popParams['validzRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'zRange' : [30,60]}     # add dict with params for this pop
-            self.paramsMap["pop"]["zRangeParamsTest"].append(params)
-
-            params = ParamsObj()
-            params.netParams.sizeZ = 70   # max size for network
-            params.netParams.popParams['invalidzRangeParams'] = {'cellType': 'PYR', 'cellModel': 'HH', 'density' : 0.8, 'numCells': 50, 'zRange' : [40,90]}     # add dict with params for this pop
-            self.paramsMap["pop"]["zRangeParamsTest"].append(params)
-
-            #net params test
-            self.paramsMap["net"]["sizeXParamsTest"] = []
-
-            params = ParamsObj()
-            params.netParams.sizeX = 70   # max size for network
-            self.paramsMap["net"]["sizeXParamsTest"].append(params)
-
-            params = ParamsObj()
-            params.netParams.sizeX = 3.53   # max size for network
-            self.paramsMap["net"]["sizeXParamsTest"].append(params)
-
-            params = ParamsObj()
-            params.netParams.sizeX = -44   # max size for network
-            self.paramsMap["net"]["sizeXParamsTest"].append(params)
-
-            self.paramsMap["net"]["shapeTest"] = []
-
-            params = ParamsObj()
-            params.netParams.shape = "cuboid"   # max size for network
-            self.paramsMap["net"]["shapeTest"].append(params)
-
-            params = ParamsObj()
-            params.netParams.shape = "ellipsoid"   # max size for network
-            self.paramsMap["net"]["shapeTest"].append(params)
-
-            params = ParamsObj()
-            params.netParams.shape = "cylinder"   # max size for network
-            self.paramsMap["net"]["shapeTest"].append(params)
-
-            params = ParamsObj()
-            params.netParams.shape = "sphere"   # max size for network
-            self.paramsMap["net"]["shapeTest"].append(params)
+            # #net params test
+            # self.paramsMap["net"]["sizeXParamsTest"] = []
+            #
+            # params = ParamsObj()
+            # params.netParams.sizeX = 70   # max size for network
+            # self.paramsMap["net"]["sizeXParamsTest"].append(params)
+            #
+            # params = ParamsObj()
+            # params.netParams.sizeX = 3.53   # max size for network
+            # self.paramsMap["net"]["sizeXParamsTest"].append(params)
+            #
+            # params = ParamsObj()
+            # params.netParams.sizeX = -44   # max size for network
+            # self.paramsMap["net"]["sizeXParamsTest"].append(params)
+            #
+            # self.paramsMap["net"]["shapeTest"] = []
+            #
+            # params = ParamsObj()
+            # params.netParams.shape = "cuboid"   # max size for network
+            # self.paramsMap["net"]["shapeTest"].append(params)
+            #
+            # params = ParamsObj()
+            # params.netParams.shape = "ellipsoid"   # max size for network
+            # self.paramsMap["net"]["shapeTest"].append(params)
+            #
+            # params = ParamsObj()
+            # params.netParams.shape = "cylinder"   # max size for network
+            # self.paramsMap["net"]["shapeTest"].append(params)
+            #
+            # params = ParamsObj()
+            # params.netParams.shape = "sphere"   # max size for network
+            # self.paramsMap["net"]["shapeTest"].append(params)
 
             #
             # # cell params test
@@ -1902,8 +1944,8 @@ class RunNetPyneTests():
 
         def runTestsWithParams(self):
 
-            #self.runPopTestsWithParams()
-            self.runNetTestsWithParams()
+            self.runPopTestsWithParams()
+            #self.runNetTestsWithParams()
             #self.runCellTestsWithParams()
             #self.runConnTestsWithParams()
             #self.runStimSourceTests()