+++ CHANGES.md
@@ -1,3 +1,11 @@
+# Version 0.7.1
+
+- Added warning when point process mechanism (eg. VecStim) required for pop is not found
+
+- Fixed bug: removed 'propList' param in list - was replaced by 'label'
+
+- Fixed bug distributing cells of population with cellList (issue #235)
+
 # Version 0.7.0
 
 IMPORTANT NOTE: NO BACKWARD REPRODUCIBILITY -- due to improvements in the random number generators it won't be possible to reproduce results of previous versions that required random values, e.g. random spike streams, connections or cell locations.